fix(test): address review feedback on #945
- Move ha_get_automation_traces test into existing test_traces.py
- Remove standalone test_traces_negative.py (new file was wrong)
- Remove @pytest.mark.asyncio (asyncio_mode = auto)
- Add type hints: mcp_client: Any, -> None
- Use result directly (no data wrapper on safe_call_tool results)
- Shorten docstrings to one sentence each
